zebra: fix locator/format inconsistencies
When configuring a locator with a node-length of 34 bits,
if 'behavior usid' is configured, then the node-length of
the locator is modified.

This change is due to the format introduction: the 'behavior'
command mimics the 'format' command and forces the locator
to take the 'usid-f3216' format in the locator settings, whereas
the configured locator settings should not be modified, and it
should still be possible to use

This commit does the following:
- it separates 'behavior usid' config from the 'format' config
- it prevents from modifying the locator settings, when format
is modified
- it invalidates the locator state when there is a config
inconsistency. The format command must be configured.

Signed-off-by: Philippe Guibert <philippe.guibert@6wind.com>
Signed-off-by: Dmytro Shytyi <dmytro.shytyi@6wind.com>